request in filters's component
